extract primitive factory
