+++ core/src/main/java/org/jdbi/v3/core/mapper/ColumnMappers.java
@@ -36,6 +36,7 @@ public class ColumnMappers implements JdbiConfig<ColumnMappers> {
     public ColumnMappers() {
         factories.add(new EnumMapperFactory());
         factories.add(new OptionalMapperFactory());
+        factories.add(new PrimitiveMapperFactory());
         factories.add(new BuiltInMapperFactory());
         factories.add(new SqlArrayMapperFactory());
     }
